Split consumer e2e tests into separate files for parallel execution
  Break up the single consumer.spec.ts (53 tests) into 6 separate files:
  - navigation.spec.ts: Homepage, All Datasets, Topic Browsing (10 tests)
  - dataset-view.spec.ts: Dataset View tab tests (5 tests)
  - dataset-filtering.spec.ts: Filtering, Pagination, Sorting (19 tests)
  - dataset-download.spec.ts: Download Panel, Download with Filters (9 tests)
  - search.spec.ts: Search functionality (5 tests)
  - feedback.spec.ts: Feedback Form (6 tests)

  This allows Playwright to run tests in parallel across 5 workers,
  reducing test execution time significantly.
