+++ tests-e2e/consumer/dataset-filtering.spec.ts
@@ -1,0 +1,209 @@
+import { test, expect } from '@playwright/test';
+
+test.describe('Dataset Filtering', () => {
+  test.describe('Filter Sidebar', () => {
+    test('Shows filter sidebar with apply button', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      // Filter apply button should be visible
+      await expect(page.getByRole('button', { name: 'Apply' })).toBeVisible();
+    });
+
+    test('Shows page size selector with options', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      const pageSizeSelect = page.locator('#page_size');
+      await expect(pageSizeSelect).toBeVisible();
+      // Verify some page size options exist
+      await expect(pageSizeSelect.locator('option')).toHaveCount(7); // 5, 10, 25, 50, 100, 250, 500
+    });
+
+    test('Shows filter groups for dataset dimensions', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      // Filter groups are rendered as details elements or divs with filter- prefix ids
+      const filterGroups = page.locator('[id^="filter-"]');
+      // Should have at least one filter group (datasets have dimensions)
+      await expect(filterGroups.first()).toBeVisible();
+    });
+
+    test('Filter checkboxes are interactive', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      // Find a filter checkbox
+      const filterCheckbox = page.locator('.checkboxes__input__filter').first();
+      if (await filterCheckbox.isVisible()) {
+        // Checkbox should be checkable
+        await filterCheckbox.check({ force: true });
+        await expect(filterCheckbox).toBeChecked();
+      }
+    });
+  });
+
+  test.describe('Page Size', () => {
+    test('Can change page size to 25', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      await page.locator('#page_size').selectOption('25');
+      await page.getByRole('button', { name: 'Apply' }).click();
+      await expect(page).toHaveURL(/page_size=25/);
+    });
+
+    test('Can change page size to 50', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      await page.locator('#page_size').selectOption('50');
+      await page.getByRole('button', { name: 'Apply' }).click();
+      await expect(page).toHaveURL(/page_size=50/);
+    });
+
+    test('Page size persists after applying filters', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      await page.locator('#page_size').selectOption('25');
+      await page.getByRole('button', { name: 'Apply' }).click();
+      // After redirect, page size should still be selected
+      await expect(page.locator('#page_size')).toHaveValue('25');
+    });
+  });
+
+  test.describe('Applying Filters', () => {
+    test('Applying filters redirects to filtered URL', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      // Apply filters (even without changing anything, it should work)
+      await page.getByRole('button', { name: 'Apply' }).click();
+      // URL should change to include /filtered/ path
+      await expect(page).toHaveURL(/\/filtered\//);
+    });
+
+    test('Filtered view shows data table', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      await page.getByRole('button', { name: 'Apply' }).click();
+      // Table should still be visible in filtered view
+      await expect(page.locator('table')).toBeVisible();
+    });
+
+    test('Filter selection is preserved in filtered view', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      // Select a specific page size before filtering
+      await page.locator('#page_size').selectOption('50');
+      await page.getByRole('button', { name: 'Apply' }).click();
+      // Page size should be preserved
+      await expect(page.locator('#page_size')).toHaveValue('50');
+    });
+  });
+
+  test.describe('Filter Controls', () => {
+    test('Select all and None links are available in filter groups', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      // Filter groups have "Select all" and "None" controls
+      const selectAllLink = page.locator('[data-action="select-all"]').first();
+      const clearLink = page.locator('[data-action="clear"]').first();
+      await expect(selectAllLink).toBeVisible();
+      await expect(clearLink).toBeVisible();
+    });
+
+    test('Not filtered checkbox is checked by default', async ({ page }) => {
+      await page.goto('/en-GB/all');
+      const firstDataset = page.locator('.index-list__item a').first();
+      await firstDataset.click();
+      // The "Not filtered" checkbox (input with id ending in "-all") should be checked by default
+      const notFilteredCheckbox = page.locator('input[type="checkbox"][id$="-all"]').first();
+      await expect(notFilteredCheckbox).toBeChecked();
+    });
+  });
+});
+
+test.describe('Dataset Pagination', () => {
+  test('Shows pagination info when data has multiple pages', async ({ page }) => {
+    await page.goto('/en-GB/all');
+    const firstDataset = page.locator('.index-list__item a').first();
+    await firstDataset.click();
+    // Set small page size to ensure pagination
+    await page.locator('#page_size').selectOption('5');
+    await page.getByRole('button', { name: 'Apply' }).click();
+    // Pagination may or may not appear depending on data size
+    // Just verify the page loaded successfully with the filtered view
+    await expect(page.locator('table')).toBeVisible();
+  });
+
+  test('Can navigate to next page if available', async ({ page }) => {
+    await page.goto('/en-GB/all');
+    const firstDataset = page.locator('.index-list__item a').first();
+    await firstDataset.click();
+    await page.locator('#page_size').selectOption('5');
+    await page.getByRole('button', { name: 'Apply' }).click();
+    // Check if next page link exists
+    const nextLink = page.getByRole('link', { name: /next/i });
+    if (await nextLink.isVisible()) {
+      await nextLink.click();
+      await expect(page).toHaveURL(/page_number=2/);
+    }
+  });
+
+  test('Page number is preserved in URL', async ({ page }) => {
+    await page.goto('/en-GB/all');
+    const firstDataset = page.locator('.index-list__item a').first();
+    await firstDataset.click();
+    await page.locator('#page_size').selectOption('5');
+    await page.getByRole('button', { name: 'Apply' }).click();
+    const nextLink = page.getByRole('link', { name: /next/i });
+    if (await nextLink.isVisible()) {
+      await nextLink.click();
+      // URL should contain page_number parameter
+      await expect(page).toHaveURL(/page_number/);
+    }
+  });
+});
+
+test.describe('Dataset Table Sorting', () => {
+  test('Table headers are displayed', async ({ page }) => {
+    await page.goto('/en-GB/all');
+    const firstDataset = page.locator('.index-list__item a').first();
+    await firstDataset.click();
+    // Table should have header row
+    const tableHeaders = page.locator('table thead th, table th');
+    await expect(tableHeaders.first()).toBeVisible();
+  });
+
+  test('Sortable columns have sort links', async ({ page }) => {
+    await page.goto('/en-GB/all');
+    const firstDataset = page.locator('.index-list__item a').first();
+    await firstDataset.click();
+    // Look for sort links/buttons in table headers
+    const sortableHeader = page.locator('th a, th button').first();
+    if (await sortableHeader.isVisible()) {
+      await expect(sortableHeader).toBeVisible();
+    }
+  });
+
+  test('Clicking sort changes URL parameters', async ({ page }) => {
+    await page.goto('/en-GB/all');
+    const firstDataset = page.locator('.index-list__item a').first();
+    await firstDataset.click();
+    // Apply filters first to get to filtered view
+    await page.getByRole('button', { name: 'Apply' }).click();
+    // Look for sortable column link
+    const sortLink = page.locator('th a').first();
+    if (await sortLink.isVisible()) {
+      await sortLink.click();
+      // URL should contain sort parameters
+      await expect(page).toHaveURL(/sort_by/);
+    }
+  });
+});
